refactor(combo): extract shadow routing to combo/shadowRouting.ts
